requirements.txt
python_dotenv
typer
click
youtube-transcript-api
ollama
requests
lxml
beautifulsoup4